@@ -1,74 +1,92 @@
 {
-  "name": "hyperledger-explorer",
-  "version": "0.3.9",
-  "description": "hyperledger-explorer",
-  "private": true,
-  "main": "main.js",
-  "keywords": [
-    "fabric-client sample app",
-    "v1.4.0 fabric nodesdk sample"
-  ],
-  "engines": {
-    "node": "^8.11.2",
-    "npm": "^5.10.0"
-  },
-  "dependencies": {
-    "ajv": "^5.5.2",
-    "app-root-path": "^2.0.1",
-    "asn1.js": "^5.0.1",
-    "body-parser": "^1.17.2",
-    "child_process": "^1.0.2",
-    "classnames": "^2.2.6",
-    "co": "^4.6.0",
-    "compression": "^1.7.2",
-    "convert-hex": "^0.1.0",
-    "ejs": "^2.5.6",
-    "enum": "^2.5.0",
-    "express": "^4.15.3",
-    "express-fileupload": "^0.4.0",
-    "fabric-ca-client": "^1.4.0",
-    "fabric-client": "^1.4.0",
-    "fabric-network": "^1.4.0",
-    "fast-stats": "0.0.3",
-    "fs-extra": "^6.0.1",
-    "http-status-codes": "^1.3.0",
-    "js-sha256": "^0.9.0",
-    "log4js": "^0.6.38",
-    "multer": "^1.3.0",
-    "pg": "^7.4.1",
-    "prettyjson": "^1.2.1",
-    "prop-types": "^15.6.2",
-    "stomp-broker-js": "^0.1.3",
-    "stompjs": "^2.3.3",
-    "swagger-ui-express": "^3.0.9",
-    "utf8-bytes": "0.0.1",
-    "winston": "^2.2.0",
-    "ws": "^5.1.1"
-  },
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "babel-eslint": "^9.0.0",
-    "eslint": "5.12.0",
-    "eslint-config-airbnb": "^17.0.0",
-    "eslint-plugin-import": "^2.13.0",
-    "eslint-plugin-jsx-a11y": "^6.1.0",
-    "eslint-plugin-react": "^7.10.0",
-    "husky": "^0.14.3",
-    "lint-staged": "^7.2.0",
-    "prettier": "^1.13.7"
-  },
-  "lint-staged": {
-    "**/*.{js,jsx,json,css}": [
-      "prettier --single-quote --write",
-      "git add"
-    ]
-  },
-  "scripts": {
-    "docker_build": "docker build --squash -t $npm_package_name .",
-    "docker_push": "docker push ${DOCKER_REGISTRY}/$npm_package_name:$npm_package_version",
-    "docker_tag": "docker tag $npm_package_name ${DOCKER_REGISTRY}/$npm_package_name:$npm_package_version",
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "precommit": "lint-staged"
-  }
+	"name": "hyperledger-explorer",
+	"version": "0.3.9",
+	"description": "hyperledger-explorer",
+	"private": true,
+	"main": "main.js",
+	"keywords": [
+		"fabric-client sample app",
+		"v1.4.0 fabric nodesdk sample"
+	],
+	"engines": {
+		"node": "^8.11.2",
+		"npm": "^5.10.0"
+	},
+	"dependencies": {
+		"ajv": "^5.5.2",
+		"app-root-path": "^2.0.1",
+		"asn1.js": "^5.0.1",
+		"body-parser": "^1.17.2",
+		"child_process": "^1.0.2",
+		"classnames": "^2.2.6",
+		"co": "^4.6.0",
+		"compression": "^1.7.2",
+		"convert-hex": "^0.1.0",
+		"ejs": "^2.5.6",
+		"enum": "^2.5.0",
+		"express": "^4.15.3",
+		"express-fileupload": "^0.4.0",
+		"fabric-ca-client": "^1.4.0",
+		"fabric-client": "^1.4.0",
+		"fabric-network": "^1.4.0",
+		"fast-stats": "0.0.3",
+		"fs-extra": "^6.0.1",
+		"grpc": "^1.20.3",
+		"http-status-codes": "^1.3.0",
+		"js-sha256": "^0.9.0",
+		"jsonwebtoken": "^8.5.0",
+		"log4js": "^0.6.38",
+		"multer": "^1.3.0",
+		"npm-run-all": "^4.1.5",
+		"passport": "^0.4.0",
+		"passport-local": "^1.0.0",
+		"pg": "^7.4.1",
+		"prettyjson": "^1.2.1",
+		"prop-types": "^15.6.2",
+		"save": "^2.3.3",
+		"stomp-broker-js": "^0.1.3",
+		"stompjs": "^2.3.3",
+		"swagger-ui-express": "^3.0.9",
+		"utf8-bytes": "0.0.1",
+		"winston": "^2.2.0",
+		"ws": "^5.1.1"
+	},
+	"license": "Apache-2.0",
+	"devDependencies": {
+		"babel-eslint": "10.0.1",
+		"eslint": "^5.16.0",
+		"eslint-config-airbnb": "^17.0.0",
+		"eslint-plugin-import": "^2.13.0",
+		"eslint-plugin-jsx-a11y": "^6.1.0",
+		"eslint-plugin-react": "^7.10.0",
+		"eslint-plugin-spellcheck": "0.0.11",
+		"husky": "^0.14.3",
+		"lint-staged": "^7.2.0",
+		"prettier": "^1.13.7"
+	},
+	"lint-staged": {
+		"**/*.{js,jsx,json,css}": [
+			"prettier --single-quote --write",
+			"git add"
+		],
+		"**/*.{js,jsx,css}": [
+			"eslint",
+			"git add"
+		]
+	},
+	"scripts": {
+		"docker_build": "docker build --squash -t $npm_package_name .",
+		"docker_push": "docker push ${DOCKER_REGISTRY}/$npm_package_name:$npm_package_version",
+		"docker_tag": "docker tag $npm_package_name ${DOCKER_REGISTRY}/$npm_package_name:$npm_package_version",
+		"lint": "eslint .",
+		"linter": "eslint",
+		"lint:fix": "eslint . --fix",
+		"precommit": "lint-staged && ./verify-license.sh",
+		"e2e-test-check-tool": "/bin/bash -c 'if [[ -z $(which configtxgen) ]]; then echo \"### Need to install tools ###\n\"; exit -1; fi'",
+		"e2e-test-check-img": "/bin/bash -c 'if [[ -z $(docker images -q hyperledger/fabric-peer:latest) ]]; then echo \"### Need to pull fabric images ###\n\"; exit -1; fi'",
+		"e2e-test-setup-env": "cd app/platform/fabric/e2e-test; if [ ! -e e2e-test ]; then virtualenv e2e-test; fi && . ./e2e-test/bin/activate && pip install -r requirement.txt",
+		"e2e-test-setup-img": "./build_docker_image.sh",
+		"e2e-test-run": "cd app/platform/fabric/e2e-test; . ./e2e-test/bin/activate && behave explorer.feature",
+		"e2e-test": "run-s e2e-test-check-tool e2e-test-check-img e2e-test-setup-env e2e-test-setup-img e2e-test-run"
+	}
 }
