@@ -1,6 +1,6 @@
 /*
-    SPDX-License-Identifier: Apache-2.0
-*/
+ * SPDX-License-Identifier: Apache-2.0
+ */
 
 const syncconfig = require('./explorerconfig.json');
 const helper = require('./common/helper');
@@ -14,69 +14,76 @@ const ExplorerSender = require('./sync/sender/ExplorerSender');
 const explorer_const = require('./common/ExplorerConst').explorer.const;
 const explorer_error = require('./common/ExplorerMessage').explorer.error;
 
-let syncScanner;
-
+/**
+ *
+ *
+ * @class Synchronizer
+ */
 class Synchronizer {
-  constructor(args) {
-    this.args = args;
-    this.persistence;
-    this.platform;
-  }
+	/**
+	 * Creates an instance of Synchronizer.
+	 * @param {*} args
+	 * @memberof Synchronizer
+	 */
+	constructor(args) {
+		this.args = args;
+		this.persistence = null;
+		this.platform = null;
+	}
 
-  async initialize() {
-    if (!syncconfig[explorer_const.PERSISTENCE]) {
-      throw new ExplorerError(explorer_error.ERROR_1001);
-    }
-    if (!syncconfig[syncconfig[explorer_const.PERSISTENCE]]) {
-      throw new ExplorerError(
-        explorer_error.ERROR_1002,
-        syncconfig[explorer_const.PERSISTENCE]
-      );
-    }
+	/**
+	 *
+	 *
+	 * @memberof Synchronizer
+	 */
+	async initialize() {
+		if (!syncconfig[explorer_const.PERSISTENCE]) {
+			throw new ExplorerError(explorer_error.ERROR_1001);
+		}
+		if (!syncconfig[syncconfig[explorer_const.PERSISTENCE]]) {
+			throw new ExplorerError(
+				explorer_error.ERROR_1002,
+				syncconfig[explorer_const.PERSISTENCE]
+			);
+		}
 
-    let pltfrm;
-    if (syncconfig && syncconfig.sync && syncconfig.sync.platform) {
-      pltfrm = syncconfig.sync.platform;
-    } else {
-      throw new ExplorerError(explorer_error.ERROR_1006);
-    }
+		let pltfrm;
+		if (syncconfig && syncconfig.sync && syncconfig.sync.platform) {
+			pltfrm = syncconfig.sync.platform;
+		} else {
+			throw new ExplorerError(explorer_error.ERROR_1006);
+		}
 
-    // if (!this.args || this.args.length == 0) {
-    // throw new ExplorerError(explorer_error.ERROR_1007);
-    // }
+		this.persistence = await PersistenceFactory.create(
+			syncconfig[explorer_const.PERSISTENCE],
+			syncconfig[syncconfig[explorer_const.PERSISTENCE]]
+		);
 
-    if (
-      !(this.args && this.args.length > 2 && this.args[2] === '1')
-      && syncconfig.sync.type !== explorer_const.SYNC_TYPE_HOST
-    ) {
-      throw new ExplorerError(explorer_error.ERROR_1008);
-    }
+		const sender = new ExplorerSender(syncconfig.sync);
+		sender.initialize();
+		logger.debug(' Synchronizer initialized');
+		this.platform = await SyncBuilder.build(pltfrm, this.persistence, sender);
 
-    this.persistence = await PersistenceFactory.create(
-      syncconfig[explorer_const.PERSISTENCE],
-      syncconfig[syncconfig[explorer_const.PERSISTENCE]]
-    );
+		this.platform.setPersistenceService();
 
-    const sender = new ExplorerSender(syncconfig.sync);
-    sender.initialize();
+		this.platform.setBlocksSyncTime(syncconfig.sync.blocksSyncTime);
 
-    this.platform = await SyncBuilder.build(pltfrm, this.persistence, sender);
+		await this.platform.initialize(this.args);
+	}
 
-    this.platform.setPersistenceService();
-
-    this.platform.setBlocksSyncTime(syncconfig.sync.blocksSyncTime);
-
-    await this.platform.initialize(this.args);
-  }
-
-  close() {
-    if (this.persistence) {
-      // this.persistence.closeconnection();
-    }
-    if (this.platform) {
-      this.platform.destroy();
-    }
-  }
+	/**
+	 *
+	 *
+	 * @memberof Synchronizer
+	 */
+	close() {
+		if (this.persistence) {
+			this.persistence.closeconnection();
+		}
+		if (this.platform) {
+			this.platform.destroy();
+		}
+	}
 }
 
 module.exports = Synchronizer;
