@@ -1,14 +1,19 @@
 /*
-    SPDX-License-Identifier: Apache-2.0
-*/
+ * SPDX-License-Identifier: Apache-2.0
+ */
 
+/**
+ * Explorer constants
+ */
 exports.explorer = {
-  const: {
-    PERSISTENCE: 'persistence',
-    PLATFORMS: 'platforms',
-    PLATFORM_FABRIC: 'fabric',
-    PERSISTENCE_POSTGRESQL: 'postgreSQL',
-    SYNC_TYPE_LOCAL: 'local',
-    SYNC_TYPE_HOST: 'host'
-  }
+	const: {
+		PERSISTENCE: 'persistence',
+		PLATFORMS: 'platforms',
+		PLATFORM_FABRIC: 'fabric',
+		PERSISTENCE_POSTGRESQL: 'postgreSQL',
+		SYNC_TYPE_LOCAL: 'local',
+		SYNC_TYPE_HOST: 'host',
+		HEALTHZ: 'healthz',
+		METRICS: 'metrics'
+	}
 };
