@@ -1,6 +1,6 @@
 /*
-    SPDX-License-Identifier: Apache-2.0
-*/
+ * SPDX-License-Identifier: Apache-2.0
+ */
 
 const path = require('path');
 const fs = require('fs-extra');
@@ -22,155 +22,182 @@ const explorer_mess = require('../../../common/ExplorerMessage').explorer;
 
 const config_path = path.resolve(__dirname, '../config.json');
 
+/**
+ *
+ *
+ * @class SyncPlatform
+ */
 class SyncPlatform {
-  constructor(persistence, sender) {
-    this.network_name;
-    this.client_name;
-    this.client;
-    this.eventHub;
-    this.sender = sender;
-    this.persistence = persistence;
-    this.syncService = new SyncService(this, this.persistence);
-    this.blocksSyncTime = 60000;
-    this.client_configs;
-  }
+	/**
+	 * Creates an instance of SyncPlatform.
+	 * @param {*} persistence
+	 * @param {*} sender
+	 * @memberof SyncPlatform
+	 */
+	constructor(persistence, sender) {
+		this.network_name = null;
+		this.client_name = null;
+		this.client = null;
+		this.eventHub = null;
+		this.sender = sender;
+		this.persistence = persistence;
+		this.syncService = new SyncService(this, this.persistence);
+		this.blocksSyncTime = 60000;
+		this.client_configs = null;
+	}
 
-  async initialize(args) {
-    const _self = this;
+	/**
+	 *
+	 *
+	 * @param {*} args
+	 * @returns
+	 * @memberof SyncPlatform
+	 */
+	async initialize(args) {
+		const _self = this;
 
-    logger.debug(
-      '******* Initialization started for child client process %s ******',
-      this.client_name
-    );
+		logger.debug(
+			'******* Initialization started for child client process %s ******',
+			this.client_name
+		);
 
-    setTimeout(() => {
-      this.initialize(args);
-    }, 60000);
+		setTimeout(() => {
+			console.log('SyncPlatform initialize()-- @ ', new Date().toDateString());
+			this.initialize(args);
+		}, 30000);
 
-    // loading the config.json
-    const all_config = JSON.parse(fs.readFileSync(config_path, 'utf8'));
-    const network_configs = all_config[fabric_const.NETWORK_CONFIGS];
+		// Loading the config.json
+		const all_config = JSON.parse(fs.readFileSync(config_path, 'utf8'));
+		const network_configs = all_config[fabric_const.NETWORK_CONFIGS];
 
-    if (args.length == 0) {
-      // get the first network and first client
-      this.network_name = Object.keys(network_configs)[0];
-      this.client_name = Object.keys(
-        network_configs[Object.keys(network_configs)[0]].clients
-      )[0];
-    } else if (args.length == 1) {
-      // get the first client with respect to the passed network name
-      this.network_name = args[0];
-      this.client_name = Object.keys(
-        network_configs[this.network_name].clients
-      )[0];
-    } else {
-      this.network_name = args[0];
-      this.client_name = args[1];
-    }
+		if (args.length === 0) {
+			// Get the first network and first client
+			this.network_name = Object.keys(network_configs)[0];
+			this.client_name = network_configs[this.network_name].name;
+		} else if (args.length === 1) {
+			// Get the first client with respect to the passed network name
+			this.network_name = args[0];
+			this.client_name = Object.keys(
+				network_configs[this.network_name].clients
+			)[0];
+		} else {
+			this.network_name = args[0];
+			this.client_name = args[1];
+		}
 
-    console.log(
-      `\n${explorer_mess.message.MESSAGE_1002}`,
-      this.network_name,
-      this.client_name
-    );
+		console.log(
+			`\n${explorer_mess.message.MESSAGE_1002}`,
+			this.network_name,
+			this.client_name
+		);
 
-    // setting the block synch interval time
-    await this.setBlocksSyncTime(all_config);
+		// Setting the block synch interval time
+		await this.setBlocksSyncTime(all_config);
 
-    logger.debug('Blocks synch interval time >> %s', this.blocksSyncTime);
-    // update the discovery-cache-life as block synch interval time in global config
-    global.hfc.config.set('discovery-cache-life', this.blocksSyncTime);
-    global.hfc.config.set('initialize-with-discovery', true);
+		logger.debug('Blocks synch interval time >> %s', this.blocksSyncTime);
+		// Update the discovery-cache-life as block synch interval time in global config
+		global.hfc.config.set('discovery-cache-life', this.blocksSyncTime);
+		global.hfc.config.set('initialize-with-discovery', true);
 
-    const client_configs = network_configs[this.network_name];
+		const client_configs = network_configs[this.network_name];
 
-    this.client_configs = await FabricUtils.setOrgEnrolmentPath(client_configs);
+		this.client_configs = await FabricUtils.setOrgEnrolmentPath(client_configs);
 
-    this.client = await FabricUtils.createFabricClient(
-      this.client_configs,
-      this.client_name
-    );
-    if (!this.client) {
-      throw new ExplorerError(explorer_mess.error.ERROR_2011);
-    }
-    const peer = {
-      requests: this.client.getDefaultPeer().getUrl(),
-      mspid: this.client_configs.organizations[
-        this.client_configs.clients[this.client_name].organization
-      ].mspid
-    };
+		this.client = await FabricUtils.createFabricClient(
+			this.client_configs,
+			this.client_name
+		);
+		if (!this.client) {
+			throw new ExplorerError(explorer_mess.error.ERROR_2011);
+		}
 
-    const peerStatus = await this.client.getPeerStatus(peer);
+		// Updating the client network and other details to DB
+		const res = await this.syncService.synchNetworkConfigToDB(this.client);
+		if (!res) {
+			return;
+		}
 
-    if (peerStatus.status) {
-      // updating the client network and other details to DB
-      const res = await this.syncService.synchNetworkConfigToDB(this.client);
-      if (!res) {
-        return;
-      }
+		// Start event
+		this.eventHub = new FabricEvent(this.client, this.syncService);
+		await this.eventHub.initialize();
 
-      // start event
-      this.eventHub = new FabricEvent(this.client, this.syncService);
-      await this.eventHub.initialize();
+		/*
+		 * Setting interval for validating any missing block from the current client ledger
+		 * Set blocksSyncTime property in platform config.json in minutes
+		 */
+		setInterval(() => {
+			_self.isChannelEventHubConnected();
+		}, this.blocksSyncTime);
+		logger.debug(
+			'******* Initialization end for child client process %s ******',
+			this.client_name
+		);
+	}
 
-      // setting interval for validating any missing block from the current client ledger
-      // set blocksSyncTime property in platform config.json in minutes
-      setInterval(() => {
-        _self.isChannelEventHubConnected();
-      }, this.blocksSyncTime);
-      logger.debug(
-        '******* Initialization end for child client process %s ******',
-        this.client_name
-      );
-    } else {
-      throw new ExplorerError(explorer_mess.error.ERROR_1009);
-    }
-  }
+	/**
+	 *
+	 *
+	 * @memberof SyncPlatform
+	 */
+	async isChannelEventHubConnected() {
+		for (const [channel_name, channel] of this.client.getChannels().entries()) {
+			// Validate channel event is connected
+			const status = this.eventHub.isChannelEventHubConnected(channel_name);
+			if (status) {
+				await this.syncService.synchBlocks(this.client, channel);
+			} else {
+				// Channel client is not connected then it will reconnect
+				this.eventHub.connectChannelEventHub(channel_name);
+			}
+		}
+	}
 
-  async isChannelEventHubConnected() {
-    for (const [channel_name, channel] of this.client.getChannels().entries()) {
-      // validate channel event is connected
-      const status = this.eventHub.isChannelEventHubConnected(channel_name);
-      if (status) {
-        await this.syncService.synchBlocks(this.client, channel);
-      } else {
-        // channel client is not connected then it will reconnect
-        this.eventHub.connectChannelEventHub(channel_name);
-      }
-    }
-  }
+	setBlocksSyncTime(blocksSyncTime) {
+		if (blocksSyncTime) {
+			const time = parseInt(blocksSyncTime, 10);
+			if (!isNaN(time)) {
+				this.blocksSyncTime = time * 60 * 1000;
+			}
+		}
+	}
 
-  setBlocksSyncTime(blocksSyncTime) {
-    if (blocksSyncTime) {
-      const time = parseInt(blocksSyncTime, 10);
-      if (!isNaN(time)) {
-        // this.blocksSyncTime = 1 * 10 * 1000;
-        this.blocksSyncTime = time * 60 * 1000;
-      }
-    }
-  }
+	/**
+	 *
+	 *
+	 * @memberof SyncPlatform
+	 */
+	setPersistenceService() {
+		// Setting platform specific CRUDService and MetricService
+		this.persistence.setMetricService(
+			new MetricService(this.persistence.getPGService())
+		);
+		this.persistence.setCrudService(
+			new CRUDService(this.persistence.getPGService())
+		);
+	}
 
-  setPersistenceService() {
-    // setting platfrom specific CRUDService and MetricService
-    this.persistence.setMetricService(
-      new MetricService(this.persistence.getPGService())
-    );
-    this.persistence.setCrudService(
-      new CRUDService(this.persistence.getPGService())
-    );
-  }
+	/**
+	 *
+	 *
+	 * @param {*} notify
+	 * @memberof SyncPlatform
+	 */
+	send(notify) {
+		if (this.sender) {
+			this.sender.send(notify);
+		}
+	}
 
-  send(notify) {
-    if (this.sender) {
-      this.sender.send(notify);
-    }
-  }
-
-  destroy() {
-    if (this.eventHub) {
-      this.eventHub.disconnectEventHubs();
-    }
-  }
+	/**
+	 *
+	 *
+	 * @memberof SyncPlatform
+	 */
+	destroy() {
+		if (this.eventHub) {
+			this.eventHub.disconnectEventHubs();
+		}
+	}
 }
 
 module.exports = SyncPlatform;
