@@ -1,19 +1,19 @@
 /*
-    SPDX-License-Identifier: Apache-2.0
-*/
+ * SPDX-License-Identifier: Apache-2.0
+ */
 
 exports.fabric = {
-  const: {
-    NETWORK_CONFIGS: 'network-configs',
-    PERSISTENCE: 'persistence',
-    BLOCK_TYPE_CONFIG: 'CONFIG',
-    BLOCK_TYPE_ENDORSER_TRANSACTION: 'ENDORSER_TRANSACTION',
-    CHAINCODE_LSCC: 'lscc',
-    NOTITY_TYPE_NEWCHANNEL: '1',
-    NOTITY_TYPE_UPDATECHANNEL: '2',
-    NOTITY_TYPE_CHAINCODE: '3',
-    NOTITY_TYPE_BLOCK: '4',
-    NOTITY_TYPE_EXISTCHANNEL: '5',
-    NOTITY_TYPE_CLIENTERROR: '6'
-  }
+	const: {
+		NETWORK_CONFIGS: 'network-configs',
+		PERSISTENCE: 'persistence',
+		BLOCK_TYPE_CONFIG: 'CONFIG',
+		BLOCK_TYPE_ENDORSER_TRANSACTION: 'ENDORSER_TRANSACTION',
+		CHAINCODE_LSCC: 'lscc',
+		NOTITY_TYPE_NEWCHANNEL: '1',
+		NOTITY_TYPE_UPDATECHANNEL: '2',
+		NOTITY_TYPE_CHAINCODE: '3',
+		NOTITY_TYPE_BLOCK: '4',
+		NOTITY_TYPE_EXISTCHANNEL: '5',
+		NOTITY_TYPE_CLIENTERROR: '6'
+	}
 };
