@@ -1,35 +1,51 @@
 /*
-    SPDX-License-Identifier: Apache-2.0
-*/
+ * SPDX-License-Identifier: Apache-2.0
+ */
 
 const explorer_const = require('../../common/ExplorerConst').explorer.const;
 
+/**
+ *
+ *
+ * @class ExplorerSender
+ */
 class ExplorerSender {
-  constructor(syncconfig) {
-    this.syncType = syncconfig.type;
-    this.syncSenderHandler;
-  }
+	/**
+	 * Creates an instance of ExplorerSender.
+	 * @param {*} syncconfig
+	 * @memberof ExplorerSender
+	 */
+	constructor(syncconfig) {
+		this.syncType = syncconfig.type;
+		this.syncSenderHandler = null;
+	}
 
-  async initialize() {
-    if (this.syncType && this.syncType === explorer_const.SYNC_TYPE_LOCAL) {
-      const ForkSenderHandler = require('./ForkSenderHandler');
-      this.syncSenderHandler = new ForkSenderHandler();
-    }
-    if (this.syncSenderHandler) {
-      this.syncSenderHandler.initialize();
-    }
-  }
+	/**
+	 *
+	 *
+	 * @memberof ExplorerSender
+	 */
+	async initialize() {
+		if (this.syncType && this.syncType === explorer_const.SYNC_TYPE_LOCAL) {
+			const ForkSenderHandler = require('./ForkSenderHandler');
+			this.syncSenderHandler = new ForkSenderHandler();
+		}
+		if (this.syncSenderHandler) {
+			this.syncSenderHandler.initialize();
+		}
+	}
 
-  send(message) {
-    if (this.syncSenderHandler) {
-      this.syncSenderHandler.send(message);
-    }
-  }
-
-  close() {
-    if (this.syncSenderHandler) {
-    }
-  }
+	/**
+	 *
+	 *
+	 * @param {*} message
+	 * @memberof ExplorerSender
+	 */
+	send(message) {
+		if (this.syncSenderHandler) {
+			this.syncSenderHandler.send(message);
+		}
+	}
 }
 
 module.exports = ExplorerSender;
