@@ -1,14 +1,10 @@
+// SPDX-License-Identifier: Apache-2.0
 const expect = require('chai').expect;
 const assert = require('assert');
 const chai = require('chai');
 
 const should = chai.should();
 const { spy, stub } = require('sinon');
-const config = require('../../app/platform/fabric/config');
-const appconfig = require('../../appconfig.json');
-
-const host = process.env.HOST || appconfig.host;
-const port = process.env.PORT || appconfig.port;
 const sinon = require('sinon');
 const request = require('request');
 
@@ -16,39 +12,40 @@ const base = 'http://localhost:1337';
 const chaincode = require('./fixtures/chaincode.json');
 
 describe('GET /api/chaincode/:channel', () => {
-  before(() => {
-    this.get = sinon.stub(request, 'get');
-    this.post = sinon.stub(request, 'post');
-    this.put = sinon.stub(request, 'put');
-    this.delete = sinon.stub(request, 'delete');
-  });
+	before(() => {
+		this.get = sinon.stub(request, 'get');
+		this.post = sinon.stub(request, 'post');
+		this.put = sinon.stub(request, 'put');
+		this.delete = sinon.stub(request, 'delete');
+	});
 
-  after(() => {
-    request.get.restore();
-    request.post.restore();
-    request.put.restore();
-    request.delete.restore();
-  });
-  it('should return chaincode ', (done) => {
-    const obj = chaincode;
-    this.get.yields(null, JSON.stringify(obj));
-    request.get(
-      `${`${base}` + '/api/chaincode/'}${config.channel}`,
-      (err, body) => {
-        body = JSON.parse(body);
-        body.should.include.keys('status', 'chaincode');
-        body.status.should.eql(200);
-        for (let i = 0; i < body.chaincode.length; i++) {
-          body.chaincode[i].should.include.keys(
-            'channelName',
-            'version',
-            'path',
-            'chaincodename',
-            'txCount'
-          );
-        }
-        done();
-      }
-    );
-  });
+	after(() => {
+		request.get.restore();
+		request.post.restore();
+		request.put.restore();
+		request.delete.restore();
+	});
+	it('should return chaincode ', done => {
+		const obj = chaincode;
+		this.get.yields(null, JSON.stringify(obj));
+		request.get(
+			`${base}` +
+				'/api/chaincode/6571ce3234a8808327849841eb9ed43a717f7f5bf430e1fb42f922f70185404d',
+			(err, body) => {
+				body = JSON.parse(body);
+				body.should.include.keys('status', 'chaincode');
+				body.status.should.eql(200);
+				for (let i = 0; i < body.chaincode.length; i++) {
+					body.chaincode[i].should.include.keys(
+						'channelName',
+						'version',
+						'path',
+						'chaincodename',
+						'txCount'
+					);
+				}
+				done();
+			}
+		);
+	});
 });
