@@ -2,19 +2,30 @@
  *    SPDX-License-Identifier: Apache-2.0
  */
 
+import React from 'react';
+import Enzyme, { shallow, mount } from 'enzyme';
+import { unwrap } from '@material-ui/core/test-utils';
+import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
+import Adapter from 'enzyme-adapter-react-16';
+import { createMuiTheme } from '@material-ui/core/styles';
 import ChannelForm from './ChannelForm';
 
-const setup = () => {
-  const wrapper = shallow(<ChannelForm />);
+Enzyme.configure({ adapter: new Adapter() });
 
-  return {
-    wrapper,
-  };
-};
+const ComponentNaked = unwrap(ChannelForm);
 
-describe('ChannelForm', () => {
-  test('ChannelForm component should render', () => {
-    const { wrapper } = setup();
-    expect(wrapper.exists()).toBe(true);
-  });
+describe('<ChannelForm />', () => {
+	it('with shallow', () => {
+		const wrapper = shallow(<ComponentNaked classes={{}} />);
+		expect(wrapper.exists()).toBe(true);
+	});
+
+	it('with mount', () => {
+		const wrapper = mount(
+			<MuiThemeProvider theme={createMuiTheme()}>
+				<ChannelForm />
+			</MuiThemeProvider>
+		);
+		expect(wrapper.exists()).toBe(true);
+	});
 });
