@@ -3,22 +3,25 @@
  */
 
 import React from 'react';
-import Enzyme, { shallow, render, mount } from 'enzyme';
+/* eslint-disable */
+import Enzyme, { shallow, render, mount, configure } from 'enzyme';
+/* eslint-enable */
 import sinon from 'sinon';
-import { configure } from 'enzyme';
+// import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import localstorage from 'mock-local-storage';
 
 configure({ adapter: new Adapter() });
 
-global.matchMedia = global.matchMedia
-  || function () {
-    return {
-      matches: false,
-      addListener() {},
-      removeListener() {},
-    };
-  };
+global.matchMedia =
+	global.matchMedia ||
+	function() {
+		return {
+			matches: false,
+			addListener() {},
+			removeListener() {}
+		};
+	};
 
 global.React = React;
 global.shallow = shallow;
